Finished code update and tested with GitLab CE 11.2.3
I have removed the adding of skipped Mantis issues as didn't really see the point if the issues db is empty

When testing needed to run the script twice - once to add (which didn't close any closed issues) then again which ran through an updated the issues closing as needed. 

Have also commented out the categories, labels, etc. bit as couldn't make it work

Hopefully this works for someone